feat(ui): add my activities
