+++ ee/tabby-ui/lib/stores/answer-engine-store.tsx
@@ -2,10 +2,14 @@ import { create } from 'zustand'
 
 export interface AnswerEngineState {
   threadsPageNo: number
+  myThreadsPageNo: number
+  threadsTab: 'all' | 'mine'
 }
 
 const initialState: AnswerEngineState = {
-  threadsPageNo: 1
+  threadsPageNo: 1,
+  myThreadsPageNo: 1,
+  threadsTab: 'all'
 }
 
 export const useAnswerEngineStore = create<AnswerEngineState>()(() => {
@@ -20,4 +24,16 @@ export const resetThreadsPageNo = () => {
 
 export const setThreadsPageNo = (pageNo: number) => {
   return set(() => ({ threadsPageNo: pageNo }))
+}
+
+export const setMyThreadsPageNo = (pageNo: number) => {
+  return set(() => ({ myThreadsPageNo: pageNo }))
+}
+
+export const resetMyThreadsPageNo = () => {
+  return set(() => ({ myThreadsPageNo: initialState.threadsPageNo }))
+}
+
+export const setThreadsTab = (tab: 'all' | 'mine') => {
+  return set(() => ({ threadsTab: tab }))
 }
